# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Payments system",
        "title": "Payments API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/login": {
            "post": {
                "description": "Authenticates a user and signs a JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "User Login",
                "parameters": [
                    {
                        "description": "Login Credentials",
                        "name": "email\u0026password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_models.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_errs.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_errs.Response"
                        }
                    }
                }
            }
        },
        "/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Gets user`s wallet information (name, email).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Gets user's wallet information",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_errs.Response"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates user (name, password).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Updates user",
                "parameters": [
                    {
                        "description": "New name and/or password",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_models.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_errs.Response"
                        }
                    }
                }
            }
        },
        "/me/wallets": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Gets all user`s wallets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Gets all user's wallets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_models.WalletsDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_errs.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Creates wallet",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_models.WalletID"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_errs.Response"
                        }
                    }
                }
            }
        },
...
        "/me/wallets/{wallet_id}/deposit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deposits funds to a specified wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Deposits funds to a specified wallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "wallet ID",
                        "name": "wallet_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Amount to deposit",
                        "name": "amount",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_models.DepositRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_models.NewBalanceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_errs.Response"
                        }
                    }
                }
            }
        },
        "/me/wallets/{wallet_id}/transactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Gets history of transactions of an wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Gets history of transactions of an wallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "wallet ID",
                        "name": "wallet_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_models.TransactionHistory"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_errs.Response"
                        }
                    }
                }
            }
        },
        "/me/wallets/{wallet_id}/transfer": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Transfers funds to a specified wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Transfers funds to a specified wallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "wallet ID",
                        "name": "wallet_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Transfer details",
                        "name": "details",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_models.TransferRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_models.NewBalanceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_errs.Response"
                        }
                    }
                }
            }
        },
        "/me/wallets/{wallet_id}/withdraw": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Withdraws funds from a specified wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Withdraws funds from a specified wallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "wallet ID",
                        "name": "wallet_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Amount to withdraw",
                        "name": "amount",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_models.WithdrawRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_models.NewBalanceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_errs.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_SlayerSv_payments_internal_shared_errs.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "github_com_SlayerSv_payments_internal_shared_models.DepositRequest": {
            "type": "object",
            "required": [
                "amount"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                }
            }
        },
        "github_com_SlayerSv_payments_internal_shared_models.LoginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github_com_SlayerSv_payments_internal_shared_models.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "github_com_SlayerSv_payments_internal_shared_models.NewBalanceResponse": {
            "type": "object",
            "properties": {
                "new_balance": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_SlayerSv_payments_internal_shared_models.TransactionDTO": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "completed_at": {
                    "type": "string"
                },
                "donor_email": {
                    "type": "string"
                },
                "donor_name": {
                    "type": "string"
                },
                "donor_wallet_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "op_type": {
                    "type": "string"
                },
                "receiver_email": {
                    "type": "string"
                },
                "receiver_name": {
                    "type": "string"
                },
                "receiver_wallet_id": {
                    "type": "string"
                }
            }
        },
        "github_com_SlayerSv_payments_internal_shared_models.TransactionHistory": {
            "type": "object",
            "properties": {
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_models.TransactionDTO"
                    }
                }
            }
        },
        "github_com_SlayerSv_payments_internal_shared_models.TransferRequest": {
            "type": "object",
            "required": [
                "amount",
                "receiver_wallet_id"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "receiver_wallet_id": {
                    "type": "string"
                }
            }
        },
        "github_com_SlayerSv_payments_internal_shared_models.UpdateUserRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github_com_SlayerSv_payments_internal_shared_models.WalletDTO": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                }
            }
        },
        "github_com_SlayerSv_payments_internal_shared_models.WalletID": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "github_com_SlayerSv_payments_internal_shared_models.WalletsDTO": {
            "type": "object",
            "properties": {
                "wallets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_SlayerSv_payments_internal_shared_models.WalletDTO"
                    }
                }
            }
        },
        "github_com_SlayerSv_payments_internal_shared_models.WithdrawRequest": {
            "type": "object",
            "required": [
                "amount"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}